床を白塗りしてタイヤ痕を黒にしました。Rキーで床を白くすることが出来ます。

+++ Src/Engine/FHelper.cpp
@@ -196,7 +196,7 @@ bool FHelper::RayToModelCollision(RayToModelCollisionData CollisionData, Vec3& I
 		for (auto& index : hitPolygon) {
 			if (fabs(index.distance_) < fabs(minDistance)) {
 				minDistance = index.distance_;
-				min = &index - &hitPolygon[0];
+				min = static_cast<int>(&index - &hitPolygon[0]);
 			}
 		}
